requirements.txt
qdrant-client
sentence-transformers
ollama
PyPDF2
python-docx
markdown
tiktoken